requirements.txt

langchain
langchain_core
langchain-openai
fastapi
requests
python-dotenv